requirements.txt
streamlit 
torch 
timm 
opencv-python 
numpy 
pandas 
Pillow
torchvision
streamlit-option-menu